### 2025-10-15 c296cb0
Create paper.yml
