### 2025-10-15 c5e7a8b
Update paper.yml
