### 2025-10-15 daed4d7
Update paper.yml
