+++ joss-paper/.github/workflows/paper.yml
@@ -3,7 +3,7 @@ name: Build JOSS Paper
 on:
   push:
     branches: [paper]
-    paths: ['paper/**']
+    paths: ['joss-paper/paper/**']
 
 jobs:
   paper:
