### 2025-10-15 3798b9d
Update paper.md
