### 2025-10-15 6002cfb
Update paper.yml
