### 2025-10-15 ded0ff1
Update paper.md
