### 2025-10-15 7fef944
Create joss-paper.yml
