### 2025-10-15 992d06e
Delete joss-paper/.github/workflows directory
