### 2026-02-20 #10086 @editorialbot
[REVIEW]: TimeHiVE: Hierarchical Moving-Window Time Series Analysis Toolkit in R
**Submitting author:** @Ouroboro (<a href="http://orcid.org/0000-0003-2985-6358">Vladimiro Boselli</a>)
**Repository:** https://github.com/Ouroboro/TimeHiVE
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.0
