### 2026-03-02 #10086 @g-rppl
Hi @Ouroboro, this submission looks really interesting!

Unfortunately, starting the review, I came across two major issues:

- On Windows, the implementation of parallel computation is not supported by `parallel::mclapply()`, I have opened an issue in the package repo.
- It looks like JOSS has recently adjusted its paper requirements. If applicable (@mahfuz05062?), please add and adjust the required sections mentioned [here](https://github.com/openjournals/joss-reviews/issues/10086#issuecomment-3931809144).
