### 2026-03-02 #10086 @Ouroboro
> Hi [@Ouroboro](https://github.com/Ouroboro), this submission looks really interesting!
>
> Unfortunately, starting the review, I came across two major issues:
>
>     On Windows, the implementation of parallel computation is not supported by `parallel::mclapply()`, I have opened an issue in the package repo.
>
>     It looks like JOSS has recently adjusted its paper requirements. If applicable ([@mahfuz05062](https://github.com/mahfuz05062)?), please add and adjust the required sections mentioned [here](https://github.com/openjournals/joss-reviews/issues/10086#issuecomment-3931809144).

Thank you for catching this Windows compatibility issue. You are absolutely right: `parallel::mclapply()` relies on forking and is not supported on Windows, which causes the example to fail on that platform.

I will fix this by replacing all uses of `mclapply` with a cross‑platform parallel solution. Specifically, I plan to implement a small wrapper that uses `mclapply` on Unix‑like systems (preserving performance) and falls back to `parLapply` (with an appropriate cluster) on Windows. This will ensure the package works correctly on all major operating systems without changing the existing functionality.

```
UW_lapply <- function(X, FUN, ..., mc.cores = 1L) {
  if (.Platform$OS.type == "windows") {
    cl <- parallel::makeCluster(mc.cores)
    on.exit(parallel::stopCluster(cl))
    parallel::parLapply(cl, X, FUN, ...)
  } else {
    parallel::mclapply(X, FUN, ..., mc.cores = mc.cores)
  }
}
```

I will make these changes over the next days and then update the repository, unfortunately I cannot test them on Windows but the fix should solve the issue. Thank you again for the helpful review.
