### 2026-03-13 f8fadd5
Add R-CMD-check workflow for R package validation
