+++ authors.md
@@ -1,0 +1,28 @@
+# Authors and Citation
+
+## Authors
+
+- **Vladimiro Andrea Boselli**. Author, maintainer, copyright holder.
+  [](https://orcid.org/0000-0003-2985-6358)
+
+- **Paolo Tagliolato Acquaviva d'Aragona**. Author.
+  [](https://orcid.org/0000-0002-0261-313X)
+
+- **Alessandro Oggioni**. Author.
+  [](https://orcid.org/0000-0002-7997-219X)
+
+## Citation
+
+Source:
+[`inst/CITATION`](https://github.com/Ouroboro/TimeHiVE/blob/main/inst/CITATION)
+
+Vladimiro Andrea Boselli, Paolo Tagliolato, Alessandro Oggioni. (2025).
+Time-series Hierarchical Visual Explorer (v0.1.0). Zenodo.
+https://doi.org/10.5281/zenodo.15697261
+
+    @Misc{,
+      title = {Time-series Hierarchical Visual Explorer},
+      author = {Vladimiro Andrea Boselli and Paolo Tagliolato and Alessandro Oggioni},
+      year = {2025},
+      note = {R package version v0.1.0},
+    }

[…]
