### 2026-03-13 5af3dea
fix: windows compatibility for parallel::mclapply (fallback to mc.cores=1)
