+++ R/TH_plots.R
@@ -4,19 +4,19 @@
 #' series. It returns analysis on moving average and running trends. The 
 #' input is a `data.frame` created by TH_single() function.
 #'
-#' @param `result` is a `data.frame` created by to TH_single().
-#' @param `scale` a `numeric`, scaling factor for point sizes, autoset by 
+#' @param results is a `data.frame` created by to TH_single().
+#' @param scale a `numeric`, scaling factor for point sizes, autoset by 
 #' default.
-#' @param `output_file` is a `string`, optional, filename to save plot (NULL to
+#' @param output_file is a `string`, optional, filename to save plot (NULL to
 #' skip saving), NULL by default.
-#' @param `mask` a `logical` value. If TRUE, enhances main plots by using point 
+#' @param mask a `logical` value. If TRUE, enhances main plots by using point 
 #' shapes to indicate statistical significance (if available). Only applicable 
 #' for modes: "all", "trend_with_test", and "avg_with_test". For other modes, 
 #' a warning is issued and mask is ignored.
-#' @param `mode` a `character` string specifying the computation mode. One of:
+#' @param mode a `character` string specifying the computation mode. One of:
 #' "all", "avg_only", "trend_only", "avg_trend", "trend_with_test", "avg_with_test".
 #' By default it takes the mode passed as attribute by `results`.
-#' @return A `ggplot` or `gridExtra` arranged plot object.
+#' @return A `ggplot` and/or `gridExtra` arranged plot object.
 #' @export
 #' @import ggplot2
 #' @importFrom gridExtra grid.arrange arrangeGrob
