+++ R/imports.R
@@ -1,0 +1,2 @@
+#' @importFrom stats complete.cases setNames var
+NULL
