+++ R/globals.R
@@ -1,0 +1,1 @@
+utils::globalVariables(c("V1", "V2", "V3", "V4", "V5", "V6", "V7", "V8", "center"))
