+++ NAMESPACE
@@ -14,7 +14,10 @@ importFrom(gridExtra,arrangeGrob)
 importFrom(gridExtra,grid.arrange)
 importFrom(parallel,detectCores)
 importFrom(parallel,mclapply)
+importFrom(stats,complete.cases)
 importFrom(stats,cor.test)
 importFrom(stats,lm)
 importFrom(stats,pnorm)
+importFrom(stats,setNames)
 importFrom(stats,t.test)
+importFrom(stats,var)

[…]
