### 2026-03-13 #10086 @Ouroboro
Hi @g-rppl, thank you again for your valuable observation.

We have now fixed the Windows compatibility issue as discussed. Instead of a complex cross‑platform wrapper, we opted for a simpler and more robust fallback: when the OS is Windows, we set ```mc.cores = 1```, which forces ```mclapply``` to run sequentially (avoiding the fork‑related error). On Unix‑like systems, parallel execution is preserved using ```parallel::detectCores() - 1```

BEFORE
```
  results_list <- parallel::mclapply(
    1:total_windows,  # Cambiato da LEN_MAX a total_windows
    compute_window,
    mc.cores = parallel::detectCores() - 1
  )
  ```
NOW
```
if (.Platform$OS.type == "windows") {
  mc_cores <- 1
} else {
  mc_cores <- max(1, parallel::detectCores() - 1)
}

results_list <- parallel::mclapply(
  1:total_windows,
  compute_window,
  mc.cores = mc_cores
)
```

In addition, we have set up Continuous Integration following rOpenSci's best practices (https://devguide.ropensci.org/pkg_ci.html) and the ReLTER Package (https://github.com/ropensci/ReLTER/blob/main/.github/workflows/check-standard.yaml). The package is now automatically tested on Windows, macOS, and Linux at every commit. You can see the latest successful runs here: https://github.com/Ouroboro/TimeHiVE/actions

Unfortunately I could not test the code directly on Windows, but the CI run on windows-latest passed successfully (you can check the logs from the link above). Please let me know if you encounter any further issues on your Windows machine.

Thanks again for your time and constructive review!
