### 2026-03-13 cd40087
Add testthat tests for all TH_* functions
