+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.2.0
 pkgdown_sha: ~
 articles: {}
-last_built: 2026-03-13T17:28Z
+last_built: 2026-03-13T17:45Z
 urls:
   reference: https://github.com/Ouroboro/TimeHiVE/reference
   article: https://github.com/Ouroboro/TimeHiVE/articles
