### 2026-03-13 6e7569d
fix: use getOption('mc.cores') to allow single-core mode in tests
