### 2026-03-13 bb194d3
fix: avoid heavy tests on ci.
