### 2026-03-13 2615eef
fix: avoid heavy tests on ci.
