### 2026-03-13 b8af0b9
Bump version to 0.2.0 for Zenodo release
