+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: TimeHiVE
 Title: Time-series Hierarchical Visual Explorer
-Version: 0.1.0
+Version: 0.2.0
 Authors@R: c(
     person("Vladimiro Andrea", "Boselli", ,email = "boselli.v@irea.cnr.it", role = c("aut", "cre", "cph"),
       comment = c(ORCID = "0000-0003-2985-6358")),
