### 2026-03-17 #10086 @g-rppl
Hi @Ouroboro thanks for fixing!
I now was able to test the functionality and continue the review. This software seems to be very helpful for its use case! Here are some comments:

## Software
* [x] If the series has length of less than 6, the defaults will cause an opaque error (`wrong sign in 'by' argument`). Maybe return a message with the minimum length?

## Documentation
* [x] Example usage specifies data that is not available (`data`, `sdata`, `sdata2`). Better use the simulated data?
* [x] Figures in the usage documentation do not fit results of the simulated data.

## Paper
* [x] Again I am not sure whether the structure of the paper aligns with the new JOSS guidelines (@mahfuz05062)?
* [x] User @SugarDodo69 contributed to the repository but lacks from the author list, why?
* [x] Should the statement about parallel performance should be revised due to the Windows issue?
