### 2026-03-20 f84ae92
- Add length validation for short time series - Add minimum length check in TH_single, TH_coupled, and TH_tweak - Prevent opaque seq() error when n < s - Provide clear error message with guidance - Fix max() warning in TH_plots for avg_only mode - Update tests to expect new error messages
