+++ R/TH_single.R
@@ -17,6 +17,7 @@
 #' @export
 #' @importFrom parallel mclapply detectCores
 #' @importFrom stats t.test lm pnorm
+#' @importFrom TimeHiVE TH_MK_Trend
 #' @author Vladimiro Andrea Boselli, (2025) \email{boselli.v@@irea.cnr.it}
 #' @examples
 #'  \dontrun{
@@ -45,6 +46,17 @@ TH_single <- function(series, m = NULL, s = NULL, mode = "all", alpha = 0.1) {
   n <- length(series)
   if (is.null(m)) m <- if (n > 250) ceiling(n / 200) else 1
   if (is.null(s)) s <- 6 * m
+  
+  #CHECK
+  min_required_length <- s
+  if (n < min_required_length) {
+    stop(
+      "Series length (", n, ") is too short for analysis.\n",
+      "Minimum required length is ", min_required_length, 
+      " (based on s = ", s, ").\n",
+      "Consider reducing m and s parameters, or provide longer time series."
+    )
+  }
   
   # Sequences calculation
   seq_x <- seq(s, n, by = m)
