### 2026-03-20 80582b3
Restore joss-paper folder with paper files
