+++ joss-paper/paper/paper.bib
@@ -1,0 +1,77 @@
+@misc{Bos2025,
+  title = {Time-series Hierarchical Visual Explorer},
+  author = {Boselli, V. A. and Tagliolato, P. and Oggioni, A.},
+  year = {2025},
+  note = {R package version v0.1.0},
+  doi = {10.5281/zenodo.15697261}
+}
+
+@article{Bos2025b,
+  author = {Boselli, V. A. and Tagliolato, P. and Oggioni, A.},
+  title = {A software package for exploratory time series analysis of environmental data},
+  volume = {8},
+  year = {2025},
+  doi = {10.3897/aca.8.e152121},
+  publisher = {Pensoft Publishers},
+  pages = {e152121},
+  journal = {ARPHA Conference Abstracts}
+}
+
+@article{Bru2006,
+  title = {Precipitation variability and changes in the greater Alpine region over the 1800--2003 period},
+  author = {Brunetti, M. and Maugeri, M. and Nanni, T. and Auer, I. and Böhm, R. and Schöner, W.},
+  journal = {Journal of Geophysical Research: Atmospheres},
+  volume = {111},
+  number = {D11},
+  year = {2006},
+  doi = {10.1029/2005JD006674},
+  publisher = {Wiley Online Library}
+}
+
+@article{Bru2009,
+  title = {Climate variability and change in the Greater Alpine Region over the last two centuries based on multi-variable analysis},
+  author = {Brunetti, M. and Lentini, G. and Maugeri, M. and Nanni, T. and Auer, I. and Böhm, R. and Schöner, W.},
+  journal = {International Journal of Climatology},
+  volume = {29},
+  number = {15},
+  pages = {2197--2225},
+  year = {2009},
+  doi = {10.1002/joc.1857},
+  publisher = {Wiley-Blackwell}
+}
+
+@article{Kni1966,
+  title = {A computer method for calculating Kendall's tau with ungrouped data},
+  author = {Knight, W. R.},
+  journal = {Journal of the American Statistical Association},
+  volume = {61},
+  number = {314},
+  pages = {436--439},
+  year = {1966},
+  doi = {10.1080/01621459.1966.10480879},
+  publisher = {Taylor \& Francis}
+}
+
+@article{Chr2005,
+  title = {Fast algorithms for the calculation of Kendall’s $\tau$},
+  author = {Christensen, D.},
+  journal = {Computational Statistics},
+  volume = {20},
+  number = {1},
+  pages = {51--62},
+  year = {2005},
+  doi = {10.1007/bf02736122},
+  publisher = {Springer}
+}
+
+@article{Sep2025,
+  title = {Kendallknight: An R package for efficient implementation of Kendall’s correlation coefficient computation},
+  author = {Sepulveda, M. V.},
+  journal = {PLoS ONE},
+  volume = {20},
+  number = {6},
+  pages = {e0326090},
+  year = {2025},
+  doi = {10.1371/journal.pone.0326090},
+  publisher = {Public Library of Science}
+}
