+++ joss-paper/paper/paper.md
@@ -1,0 +1,80 @@
+---
+title: 'TimeHiVE: Hierarchical Moving-Window Time Series Analysis Toolkit in R'
+authors:
+  - name: Vladimiro Andrea Boselli
+    orcid: 0000-0003-2985-6358
+    affiliation: 1
+  - name: Paolo Tagliolato
+    orcid: 0000-0002-0261-313X
+    affiliation: 1
+  - name: Alessandro Oggioni
+    orcid: 0000-0002-7997-219X
+    affiliation: 1
+affiliations:
+  - name: Institute for Electromagnetic Sensing of the Environment of the National Research Council of Italy (IREA-CNR), Italy
+    index: 1
+keywords:
+  - R
+  - time series
+  - moving-window statistics
+  - Mann-Kendall
+  - correlation
+abstract: |
+  TimeHiVE is an R package designed for hierarchical moving-window statistical analysis of time series data. The package addresses a fundamental challenge in time series analysis: the selection of appropriate window sizes for moving-window calculations. Traditional moving-window approaches require users to specify a fixed window size, which can obscure important patterns if chosen inappropriately. TimeHiVE eliminates this constraint by systematically computing statistics across all possible window sizes, enabling researchers to explore the full spectrum of temporal patterns in their data.
+bibliography: paper.bib
+---
+# Summary
+
+TimeHiVE by @Bos2025 is an R package designed for hierarchical moving-window statistical analysis of time series data. The package addresses a fundamental challenge in time series analysis: the selection of appropriate window sizes for moving-window calculations. Traditional moving-window approaches require users to specify a fixed window size, which can obscure important patterns if chosen inappropriately. TimeHiVE eliminates this constraint by systematically computing statistics across all possible window sizes, enabling researchers to explore the full spectrum of temporal patterns in their data.
+
+The package provides implementations for both single time series analysis (including means, trends, and custom statistics) and coupled time series analysis (including Pearson and Mann-Kendall correlations). TimeHiVE features Mann-Kendall statistics algorithms with O(n log n) time complexity, parallel computation capabilities for unix-based systems, and customisable visualization tools. These features make it particularly valuable for environmental and climate research, where understanding phenomena across multiple timescales is essential.
+
+
+# Statement of need
+
+Moving-window statistical analysis is a fundamental technique in time series analysis, used to study the evolution of data over time in a dynamic and localized manner. This approach involves calculating statistics on subsets of consecutive data points that progressively move along the series. The technique helps identify localized trends, seasonality, and anomalies while reducing the impact of random fluctuations, highlighting meaningful patterns.
+
+The challenge of selecting an optimal window size is well-documented in time series literature. A window that is too large may obscure important details, while one that is too narrow may be overly sensitive to noise. This problem is particularly relevant in climate science, where researchers need to examine phenomena across multiple timescales. Previous work by Brunetti and colleagues [@Bru2006; @Bru2009] demonstrated the value of hierarchical moving-window approaches for climate data analysis, but implementation required custom code.
+
+TimeHiVE makes this analytical approach accessible to a broader research community by providing a well-documented, efficient implementation in R. The package offers several advantages over existing solutions:
+
+1. It eliminates the need for *a priori* window size selection
+2. It provides optimized implementations of common statistical tests
+3. It supports both single and coupled time series analysis
+4. It offers flexible visualization capabilities
+5. It allows for custom statistical functions through its extensible architecture
+
[…]
