### 2026-03-21 #10086 @Ouroboro
> Hi [@Ouroboro](https://github.com/Ouroboro) thanks for fixing! I now was able to test the functionality and continue the review. This software seems to be very helpful for its use case! Here are some comments:
> ## Software
>
>     * [ ]  If the series has length of less than 6, the defaults will cause an opaque error (`wrong sign in 'by' argument`). Maybe return a message with the minimum length?
**Corrected, I added tests too to verify it too.**
>
>
> ## Documentation
>
>     * [ ]  Example usage specifies data that is not available (`data`, `sdata`, `sdata2`). Better use the simulated data?[ ]  Figures in the usage documentation do not fit results of the simulated data.
**The figures in the usage are just random series, in some cases generated with particular criteria. Would you like me to regenerate them and create a folder with the series used to generate the images, so that it is possible to replicate the images exactly?**
>
>
> ## Paper
>
>     * [ ]  Again I am not sure whether the structure of the paper aligns with the new JOSS guidelines ([@mahfuz05062](https://github.com/mahfuz05062))?[ ]  User [@SugarDodo69](https://github.com/SugarDodo69) contributed to the repository but lacks from the author list, why?[ ]  Should the statement about parallel performance should be revised due to the Windows issue?
**SugarDodo69 is just an other account controlled by the authors. I revised the statement about pp writing it is only for unix-based systems.**
