### 2026-03-23 #10086 @g-rppl
> The figures in the usage are just random series, in some cases generated with particular criteria. Would you like me to regenerate them and create a folder with the series used to generate the images, so that it is possible to replicate the images exactly?

@Ouroboro I think it would make sense to simulate datasets and use them throughout the example section - including the figures - making it reproducible. Otherwise users get confused why their results do not fit the figures displayed.
