+++ llms.txt
@@ -17,12 +17,31 @@ of statistical tools.
 
 ![Original Time Series](img/Fig1.png)*Figure 1: Different Moving Average
 vs Original Syntetic Time Series. Color scale on y-axis is meant only to
-match colors in Fig. 2*
+match colors in Fig. 2. Source in inst/extdata/original.csv .*
 
 ![TimeHiVE Moving Average](img/Fig2.png)*Figure 2: Representation of all
 the possible moving averages for the Original Syntetic Time Series,
 moving averages analyses represented in Fig. 1 are highlighted with
-comments on the results.*
+comments on the results. Source in inst/extdata/original.csv .*
+
+``` r
+library(TimeHiVE)
+
+# Load example time series from the package data
+series <- read.csv(system.file("extdata", "original.csv", package = "TimeHiVE"),
+                    header = FALSE)[, 1]
+
+# Perform analysis
+results <- TH_single(
+  series = series,
+  m = 1,
+  s = 1,
+  mode = "avg"
+)
+
+# Visualize results
+TH_plotc(results, mode = "avg")
+```
 
 ![Original Coupled Series](img/Fig3.png)*Figure 3: Representation of two
 Time Series positively correlated for short periods but negatively
@@ -30,7 +49,8 @@ correlated for long periods. The series are built as:
 `TS1 = 40 + 2*sin(t/2) - t/20 - rand(-2/3, 2/3)` and
 `TS2 = 15 + 2*sin(t/2) + t/7 - rand(-2/3, 2/3)`. The sine component
 creates short-term positive correlation, while the `t/n` terms drive
-long-term negative correlation.*
+long-term negative correlation. Source in inst/extdata/fseries1.csv and
+inst/extdata/fseries2.csv .*
 
 ![Original Time Series](img/Fig4.png)*Figure 4: Here we show the Moving
 Correlation Analysis for the coupled Time Series represented in Fig. 4,
@@ -38,7 +58,31 @@ the first row shows the results for Pearson’s correlation coefficient
 (Top Left) and relative p-values (Top Right), the second row shows the
 same analysis with MK’s correlation coefficients (Bottom Left) and
 relative p-values (Bottom Right). The inversion of the correlation
-between short and long period of analysis is quite clear.*
+between short and long period of analysis is quite clear. Source in
+inst/extdata/fseries1.csv and inst/extdata/fseries2.csv .*
+
+``` r
+library(TimeHiVE)
+
+# Load example time series from the package data
+series1 <- read.csv(system.file("extdata", "fseries1.csv", package = "TimeHiVE"),
+                    header = FALSE)[, 1]
+series2 <- read.csv(system.file("extdata", "fseries2.csv", package = "TimeHiVE"),
+                    header = FALSE)[, 1]
+
+# Perform analysis
+results <- TH_coupled(
+  series1 = series1,
+  series2 = series2,
+  m = 2,
+  s = 6,
+  alpha = 0.05,
+  mode = "both"
+)
+
+# Visualize results
+TH_plotc(results, mask = FALSE, mode = "all")
+```
 
 FIRST ROW: Pearson’s correlation coefficient (Top Left) and relative
 p-values (Top Right). SECOND ROW: same analysis with MK’s correlation
@@ -59,13 +103,18 @@ remotes::install_github("Ouroboro/TimeHiVE")
 
 ### Basic Example
 
+The package includes example data files in `inst/extdata/`. After
+installation, you can load them using
[…]
