### 2026-04-05 #10086 @Ouroboro
@g-rppl I added a folder inst/extdata with timeseries in .csv to reproduce the results in the README.
