### 2026-04-07 #10086 @g-rppl
Great, thanks @Ouroboro! I am done with the review @mahfuz05062 - except that I'm still unsure whether the structure of the paper aligns with the new JOSS guidelines.
