### 2026-04-27 #10086 @mahfuz05062
> Great, thanks [@Ouroboro](https://github.com/Ouroboro)! I am done with the review [@mahfuz05062](https://github.com/mahfuz05062) - except that I'm still unsure whether the structure of the paper aligns with the new JOSS guidelines.

Thanks for the review and I agree with the comment about paper format / structure @Ouroboro. @g-rppl Can we follow the format mentioned in the official guidelines? Please consult these for reference: https://joss.readthedocs.io/en/latest/paper.html and https://joss.readthedocs.io/en/latest/example_paper.html
