### 2026-04-27 da33475
Align paper with JOSS requirements
