+++ joss-paper/paper/paper.md
@@ -13,7 +13,7 @@ authors:
 affiliations:
   - name: Institute for Electromagnetic Sensing of the Environment of the National Research Council of Italy (IREA-CNR), Italy
     index: 1
-keywords:
+tags:
   - R
   - time series
   - moving-window statistics
@@ -28,7 +28,6 @@ bibliography: paper.bib
 TimeHiVE by @Bos2025 is an R package designed for hierarchical moving-window statistical analysis of time series data. The package addresses a fundamental challenge in time series analysis: the selection of appropriate window sizes for moving-window calculations. Traditional moving-window approaches require users to specify a fixed window size, which can obscure important patterns if chosen inappropriately. TimeHiVE eliminates this constraint by systematically computing statistics across all possible window sizes, enabling researchers to explore the full spectrum of temporal patterns in their data.
 
 The package provides implementations for both single time series analysis (including means, trends, and custom statistics) and coupled time series analysis (including Pearson and Mann-Kendall correlations). TimeHiVE features Mann-Kendall statistics algorithms with O(n log n) time complexity, parallel computation capabilities for unix-based systems, and customisable visualization tools. These features make it particularly valuable for environmental and climate research, where understanding phenomena across multiple timescales is essential.
-
 
 # Statement of need
 
@@ -51,6 +50,22 @@ TimeHiVE implements two main analytical functions: `TH_single()` for single time
 The package includes comprehensive visualization functions (`TH_plots()`, `TH_plotc()`, and `TH_plott()`) that generate heatmap-style representations of results, with time on the x-axis and window size on the y-axis. This visualization approach, inspired by Brunetti et al. [@Bru2006; @Bru2009], enables intuitive interpretation of patterns across timescales. A recent application has been carried out on precipitation time series derived from FAIR datasets provided by the eLTER Research Infrastructure
 [eLTER-RI](https://elter-ri.eu/), offering a first example of their potential to support preliminary data exploration and quality assessment within eLTER [@Bos2025b].
 
+# State of the field
+
+The R ecosystem offers a rich variety of packages for moving‑window (rolling) time series analysis. The **zoo** package provides `rollapply()`, a general‑purpose rolling function, together with specialised routines such as `rollmean()` [@Zei2005]. **slider** offers type‑stable sliding window functions that operate on any R data type, with support for expanding and cumulative windows [@Vau2021]. **runner** is a lightweight library that enables the application of any R function to rolling windows, with full control over window size, lag, and irregular time indices [@Gur2024]. For high‑performance computing, **RcppRoll** implements fast rolling operations (mean, median, sum, standard deviation) using C++ [@Ush2024].
+
+For trend detection, the **Kendall** package implements the Mann‑Kendall trend test and Kendall rank correlation [@Mcl2011]. The **trend** package provides a broader suite of non‑parametric trend tests and change‑point detection methods, including several variants of the Mann‑Kendall test [@Poh2023].
+
+In the domain of rolling correlation, **NonParRolCor** estimates rolling window correlations between two time series, addressing the multiple testing issue via Monte Carlo simulations [@Pol2023]. **RolWinMulCor** extends this concept to the multivariate case and can produce heatmap visualisations for a band of window lengths [@Pol2020].
+
[…]
