+++ joss-paper/paper/paper.bib
@@ -75,3 +75,84 @@
   doi = {10.1371/journal.pone.0326090},
   publisher = {Public Library of Science}
 }
+
+@article{Zei2005,
+  title={zoo: S3 Infrastructure for Regular and Irregular Time Series},
+  author={Zeileis, A. and Grothendieck, G.},
+  journal={Journal of Statistical Software},
+  volume={14},
+  number={6},
+  pages={1--27},
+  year={2005},
+  doi={10.18637/jss.v014.i06}
+}
+
+@manual{Vau2021,
+  title={slider: Sliding Window Functions},
+  author={Vaughan, D.},
+  year={2021},
+  note={R package version 0.2.2},
+  url={https://CRAN.R-project.org/package=slider}
+}
+
+@manual{Gur2024,
+  title={runner: Running Operations for Vectors},
+  author={Gurgul, M.},
+  year={2024},
+  note={R package version 0.5.1},
+  url={https://CRAN.R-project.org/package=runner}
+}
+
+@manual{Ush2024,
+  title={RcppRoll: Efficient Rolling / Windowed Operations},
+  author={Ushey, K.},
+  year={2024},
+  note={R package version 0.3.2},
+  url={https://CRAN.R-project.org/package=RcppRoll}
+}
+
+@manual{Mcl2011,
+  title={Kendall: Kendall Rank Correlation and Mann-Kendall Trend Test},
+  author={McLeod, A. I.},
+  year={2011},
+  note={R package version 2.2},
+  url={https://CRAN.R-project.org/package=Kendall}
+}
+
+@manual{Poh2023,
+  title={trend: Non-Parametric Trend Tests and Change-Point Detection},
+  author={Pohlert, T.},
+  year={2023},
+  note={R package version 1.1.6},
+  url={https://CRAN.R-project.org/package=trend}
+}
+
+@article{Pol2023,
+  title={NonParRolCor: An R package for estimating rolling correlation for two regular time series},
+  author={Polanco-Mart{\'\i}nez, J. M.},
+  journal={SoftwareX},
+  volume={22},
+  pages={101353},
+  year={2023},
+  doi={10.1016/j.softx.2023.101353},
+  publisher={Elsevier}
+}
+
+@article{Pol2020,
+  title={RolWinMulCor: An R package for estimating rolling window multiple correlation in ecological time series},
+  author={Polanco-Mart{\'\i}nez, J. M.},
+  journal={Ecological Informatics},
+  volume={60},
+  pages={101163},
+  year={2020},
+  doi={10.1016/j.ecoinf.2020.101163}
+  publisher={Elsevier}
+}
+
+@inproceedings{Erm2024,
+  title={Window Size Selection in Unsupervised Time Series Analytics: A Review and Benchmark},
+  author={Ermshaus, A. and Sch{\"a}fer, P. and Leser, U.},
+  booktitle={International Conference on Data Mining},
+  year={2024},
+  doi={10.1007/978-3-031-24378-3}
+}
