### 2026-04-27 62102e3
Align paper with JOSS requirements
