+++ joss-paper/paper/paper.bib
@@ -145,7 +145,7 @@
   volume={60},
   pages={101163},
   year={2020},
-  doi={10.1016/j.ecoinf.2020.101163}
+  doi={10.1016/j.ecoinf.2020.101163},
   publisher={Elsevier}
 }
 
