### 2026-04-27 bfe3ea3
Align paper with JOSS requirements
