### 2026-04-27 #10086 @Ouroboro
@mahfuz05062 @g-rppl @iosefa the paper has been updated accordingly and the bibliography extended.
