### 2026-04-29 #10086 @mahfuz05062
> [@mahfuz05062](https://github.com/mahfuz05062) [@g-rppl](https://github.com/g-rppl) [@iosefa](https://github.com/iosefa) the paper has been updated accordingly and the bibliography extended.

Thanks @Ouroboro . @g-rppl do the new updates satisfy the question you had about the structure?
