### 2026-08-25 #10086 @iosefa
Hi @Ouroboro — the manuscript and repository are much improved since my initial review. I have three remaining minor comments before I can check off the following items.

## comment 1:
- [ ] Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

There is a minor issue with the README example associated with Figure 2. The example does not currently run: `mode = "avg"` should be `mode = "avg_only"`, and results from `TH_single()` should be passed to `TH_plots()` rather than `TH_plotc()`, from what I can tell.

Anyway, after making those corrections, the output is slightly different from the Figure 2 shown in the README, which appears to have been generated using an earlier version of the window-boundary calculation. I dont think this a significant reproducibility problem for the manuscript, but I suggest updating the README figure so that it matches the current working example.

## Comment 2:

`TH_MK_Trend()` documents its returned p-value as two-sided, but the current expression calculates a one-sided tail probability:

https://github.com/Ouroboro/TimeHiVE/blob/2bccfaa036bcd7d7148e0cc792293504b57207af/R/TH_MK_Trend.R#L126

```r
p.value <- 0.5 - abs(0.5 - pnorm(Z))
```

For example, when Z = 0, this returns 0.5, whereas a two-sided p-value should be 1. The conventional two-sided calculation would be:

```r
p.value <- 2 * pnorm(-abs(Z))
```

Please update docs or the implementation.

## Comment 3:

- [ ] Research Impact Statement (required section): Does the paper include a “Research Impact Statement” section providing evidence of realized impact (publications, external use, integrations) or credible near-term significance (benchmarks, reproducible materials, community-readiness signals)? Is the evidence presented compelling?

The Research Impact Statement is good, but I would update just this comment from lines 94-97:

  > Peer-reviewed applications demonstrated its utility for quality assessment of daily precipitation series from eLTER sites (Aigüestortes, Spain; Svartberget, Sweden) by revealing seasonal patterns and detecting post-2010 anomalies (Boselli et al., 2025a).

I am not sure that the cited work supports characterizing this as a peer-reviewed application that demonstrated the software’s utility. It appears to describe a preliminary application of the package in a conference abstract by the authors.

Otherwise, I think this is strong.
